Adds caching, adds pressure
The weather data is now cached. Only if the last request is older than
three minutes (the default time between two measurements), a new
request to the API is submitted.

The pressure measured by the indoor module is now added to the
weather data.

A few exception handler are added.